nv-requirements.txt
django-phonenumber-field
django-user-agents
phonenumbers
pillow
pip-chill
six
